+++ shared.ts
@@ -2,3 +2,11 @@ export enum SocketTypes {
   online,
   party,
 }
+export interface TexturePack {
+  id: string;
+  name: string;
+  author: string;
+  description: string;
+  hasCombined: boolean;
+  hasSeamless: boolean;
+}