+++ userscript/src/config.ts
@@ -2,7 +2,6 @@ export default {
   ver: "$$version",
   api: "https://nutils.itsmeow.cat",
   customDelimiter: "__custom",
-  packVersion: 2,
   actualGameVersion:
     document
       .querySelector<HTMLScriptElement>(`script[src*="game.js"]`)