+++ userscript/src/settings/settings.ts
@@ -3,19 +3,16 @@
 interface Settings {
   showFPS: boolean;
   texturePack: string | null;
-  customPack: string | null;
   typewriter: boolean;
   apiKey: string;
   appearOnline: boolean;
 }
 const settingsKey = "ninjaioutils";
-const packKey = "DONTEDIT_ninja_custompack";
 
 export const SETTINGS: Settings = {
   ...{
     showFPS: true,
     texturePack: null,
-    customPack: null,
     typewriter: false,
     apiKey: "",
     appearOnline: true,
@@ -23,8 +20,3 @@ export const SETTINGS: Settings = {
   ...JSON.parse(localStorage.getItem(settingsKey) || "{}"),
 };
 export const saveSettings = () => localStorage.setItem(settingsKey, JSON.stringify(SETTINGS));
-
-export const getSavedPack = () =>
-  JSON.parse(localStorage.getItem(packKey) || '["",""]') as [string, string];
-export const savePackData = (tex: string, ter: string) =>
-  localStorage.setItem(packKey, JSON.stringify([tex, ter]));